Improve import and module dependencies
